+++ src/tools/shell.ts
@@ -1,46 +1,37 @@
 /** cwd pinned to root; non-allowlisted commands throw to a UI confirm gate; spawn is `shell: false`, tokenized argv only. */
 
-import { type ChildProcess, type SpawnOptions, spawn, spawnSync } from "node:child_process";
-import { existsSync, statSync } from "node:fs";
 import * as pathMod from "node:path";
 import { addProjectShellAllowed } from "../config.js";
 import { pauseGate } from "../core/pause-gate.js";
 import type { ToolRegistry } from "../tools.js";
 import { JobRegistry } from "./jobs.js";
 import {
-  type CommandChain,
-  UnsupportedSyntaxError,
-  chainAllowed,
-  parseCommandChain,
-  runChain,
-} from "./shell-chain.js";
+  DEFAULT_MAX_OUTPUT_CHARS,
+  DEFAULT_TIMEOUT_SEC,
+  type RunCommandResult,
+  runCommand,
+} from "./shell/exec.js";
+import { isCommandAllowed } from "./shell/parse.js";
 
-/** Kill child + descendants. Windows: taskkill /T /F. Unix: SIGKILL the process group when detached, else fall back to SIGKILL on the leader. */
-export function killProcessTree(child: ChildProcess): void {
-  if (!child.pid || child.killed) return;
-  if (process.platform === "win32") {
-    try {
-      spawnSync("taskkill", ["/pid", String(child.pid), "/T", "/F"], {
-        stdio: "ignore",
-        windowsHide: true,
-      });
-      return;
-    } catch {
-      /* fall through to SIGKILL */
-    }
-  }
-  try {
-    process.kill(-child.pid, "SIGKILL");
-    return;
-  } catch {
-    /* not a process group leader — fall through */
-  }
-  try {
-    child.kill("SIGKILL");
-  } catch {
-    /* already gone */
-  }
-}
+export {
+  BUILTIN_ALLOWLIST,
+  detectShellOperator,
+  isAllowed,
+  isCommandAllowed,
+  isDqEscape,
+  tokenizeCommand,
+} from "./shell/parse.js";
+export type { ResolveExecutableOptions, RunCommandResult } from "./shell/exec.js";
+export {
+  injectPowerShellUtf8,
+  killProcessTree,
+  prepareSpawn,
+  quoteForCmdExe,
+  resolveExecutable,
+  runCommand,
+  smartDecodeOutput,
+  withUtf8Codepage,
+} from "./shell/exec.js";
 
 export interface ShellToolsOptions {
   /** Directory to run commands in. Must be an absolute path. */
@@ -53,540 +44,6 @@ export interface ShellToolsOptions {
   /** Getter form lets `editMode === "yolo"` flip mid-session without re-registering tools. */
   allowAll?: boolean | (() => boolean);
   jobs?: JobRegistry;
-}
-
-const DEFAULT_TIMEOUT_SEC = 60;
-const DEFAULT_MAX_OUTPUT_CHARS = 32_000;
-
-/** Read-only reports + test runners whose failure mode is "exit 1 with output". */
-export const BUILTIN_ALLOWLIST: ReadonlyArray<string> = [
-  // Repo inspection
-  "git status",
-  "git diff",
-  "git log",
-  "git show",
-  "git blame",
-  "git branch",
-  "git remote",
-  "git rev-parse",
-  "git config --get",
-  // Filesystem inspection
-  "ls",
-  "pwd",
-  "cat",
-  "head",
-  "tail",
-  "wc",
-  "file",
-  "tree",
-  "find",
-  "grep",
-  "rg",
-  // Language version probes
-  "node --version",
-  "node -v",
-  "npm --version",
-  "npx --version",
-  "python --version",
-  "python3 --version",
-  "cargo --version",
-  "go version",
-  "rustc --version",
-  "deno --version",
-  "bun --version",
-  // Test runners (non-destructive by convention)
-  "npm test",
-  "npm run test",
-  "npx vitest run",
-  "npx vitest",
-  "npx jest",
-  "pytest",
-  "python -m pytest",
-  "cargo test",
-  "cargo check",
-  "cargo clippy",
-  "go test",
-  "go vet",
-  "deno test",
-  "bun test",
-  // Linters / typecheckers (read-only by convention)
-  "npm run lint",
-  "npm run typecheck",
-  "npx tsc --noEmit",
-  "npx biome check",
-  "npx eslint",
-  "npx prettier --check",
-  "ruff",
-  "mypy",
-];
-
-/** Inside `"…"` only `\"` and `\\` are escapes — `\X` otherwise stays literal so Windows paths like `"C:\Users\foo\.bar"` survive tokenization. */
-export function isDqEscape(prev: string, next: string | undefined): boolean {
-  return prev === "\\" && (next === '"' || next === "\\");
-}
-
-/** No env / glob / backtick / `$(…)` expansion — prevents bypass of allowlist via concatenation. */
-export function tokenizeCommand(cmd: string): string[] {
-  const out: string[] = [];
-  let cur = "";
-  let quote: '"' | "'" | null = null;
-  for (let i = 0; i < cmd.length; i++) {
-    const ch = cmd[i]!;
-    if (quote) {
-      if (ch === quote) {
-        quote = null;
-      } else if (quote === '"' && isDqEscape(ch, cmd[i + 1])) {
-        cur += cmd[++i];
-      } else {
-        cur += ch;
-      }
-      continue;
-    }
-    if (ch === '"' || ch === "'") {
-      quote = ch;
-      continue;
-    }
-    if (ch === " " || ch === "\t") {
-      if (cur.length > 0) {
-        out.push(cur);
-        cur = "";
-      }
-      continue;
-    }
-    cur += ch;
-  }
-  if (quote) throw new Error(`unclosed ${quote} in command`);
-  if (cur.length > 0) out.push(cur);
-  return out;
-}
-
-/** Up-front detection — without it, `dir | findstr foo` quotes `|` literal and pipe silently fails. */
-export function detectShellOperator(cmd: string): string | null {
-  const opPrefix = /^(?:2>&1|&>|\|{1,2}|&{1,2}|2>{1,2}|>{1,2}|<{1,2})/;
-  let cur = "";
-  let curQuoted = false;
-  let quote: '"' | "'" | null = null;
-  const check = (): string | null => {
-    if (cur.length === 0 && !curQuoted) return null;
-    if (!curQuoted) {
-      const m = opPrefix.exec(cur);
-      if (m) return m[0] ?? null;
-    }
-    return null;
-  };
-  for (let i = 0; i < cmd.length; i++) {
-    const ch = cmd[i]!;
-    if (quote) {
-      if (ch === quote) {
-        quote = null;
-      } else if (quote === '"' && isDqEscape(ch, cmd[i + 1])) {
-        cur += cmd[++i];
-        curQuoted = true;
-      } else {
-        cur += ch;
-        curQuoted = true;
-      }
-      continue;
-    }
-    if (ch === '"' || ch === "'") {
-      quote = ch;
-      curQuoted = true;
-      continue;
-    }
-    if (ch === " " || ch === "\t") {
-      const op = check();
-      if (op) return op;
-      cur = "";
-      curQuoted = false;
-      continue;
-    }
-    cur += ch;
-  }
-  if (quote) return null; // let tokenizeCommand throw the unclosed-quote error
-  return check();
-}
-
-/** Per-prefix demotion: an otherwise-allowlisted match falls back to the confirm gate when one of these tokens appears in the tail. Issue #257: `git branch -D` skipped review. Each token also matches its `--flag=value` form. */
-const RISKY_ARGS: Readonly<Record<string, ReadonlyArray<string>>> = {
-  // Branch / remote mutation
-  "git branch": ["-d", "-D", "--delete", "-m", "-M", "--move", "-c", "-C", "--copy", "--force"],
-  "git remote": ["add", "remove", "rm", "rename", "set-url", "set-head", "prune"],
-  // `--output` writes to an arbitrary path; `--ext-diff` invokes user-config'd external programs.
-  "git diff": ["--output", "--ext-diff"],
-  "git log": ["--output"],
-  "git show": ["--output"],
-  // `-exec*` / `-ok*` are RCE; `-delete` and `-fprint*` / `-fls` write to arbitrary paths.
-  find: [
-    "-delete",
-    "-exec",
-    "-execdir",
-    "-ok",
-    "-okdir",
-    "-fprint",
-    "-fprint0",
-    "-fprintf",
-    "-fls",
-  ],
-  // `-o FILE` writes the tree to an arbitrary path.
-  tree: ["-o"],
-  // Auto-fix mutates source files.
-  "npx eslint": ["--fix", "--fix-dry-run"],
-  "npx biome check": ["--write", "--apply", "--apply-unsafe"],
-  ruff: ["--fix", "--unsafe-fixes", "format"],
-};
-
-function tailHasRisky(tail: readonly string[], risky: readonly string[]): boolean {
-  for (const a of tail) {
-    for (const r of risky) {
-      if (a === r) return true;
-      if (a.startsWith(`${r}=`)) return true;
-    }
-  }
-  return false;
-}
-
-/** Allowlist match on leading argv tokens; demoted by `RISKY_ARGS` when a destructive flag appears in the tail. */
-export function isAllowed(cmd: string, extra: readonly string[] = []): boolean {
-  let argv: string[];
-  try {
-    argv = tokenizeCommand(cmd);
-  } catch {
-    return false;
-  }
-  if (argv.length === 0) return false;
-
-  const allowlist = [...BUILTIN_ALLOWLIST, ...extra];
-  for (const prefix of allowlist) {
-    const prefixTokens = prefix.split(" ");
-    if (argv.length < prefixTokens.length) continue;
-    let match = true;
-    for (let i = 0; i < prefixTokens.length; i++) {
-      if (argv[i] !== prefixTokens[i]) {
-        match = false;
-        break;
-      }
-    }
-    if (!match) continue;
-
-    const risky = RISKY_ARGS[prefix];
-    if (risky && tailHasRisky(argv.slice(prefixTokens.length), risky)) return false;
-    return true;
-  }
-  return false;
-}
-
-/** For chain commands, every segment must individually clear the allowlist. */
-export function isCommandAllowed(cmd: string, extra: readonly string[] = []): boolean {
-  let chain: CommandChain | null;
-  try {
-    chain = parseCommandChain(cmd);
-  } catch {
-    return false;
-  }
-  if (chain === null) return isAllowed(cmd, extra);
-  return chainAllowed(chain, (seg) => isAllowed(seg, extra));
-}
-
-export interface RunCommandResult {
-  exitCode: number | null;
-  /** Combined stdout+stderr, truncated to `maxOutputChars` with a marker. */
-  output: string;
-  /** True when the process was killed for exceeding `timeoutSec`. */
-  timedOut: boolean;
-}
-
-export async function runCommand(
-  cmd: string,
-  opts: {
-    cwd: string;
-    timeoutSec?: number;
-    maxOutputChars?: number;
-    signal?: AbortSignal;
-  },
-): Promise<RunCommandResult> {
-  const timeoutSec = opts.timeoutSec ?? DEFAULT_TIMEOUT_SEC;
-  const maxChars = opts.maxOutputChars ?? DEFAULT_MAX_OUTPUT_CHARS;
-  const argv = tokenizeCommand(cmd);
-  if (argv.length === 0) throw new Error("run_command: empty command");
-  const chain = parseCommandChain(cmd);
-  if (chain !== null) {
-    return await runChain(chain, {
-      cwd: opts.cwd,
-      timeoutSec,
-      maxOutputChars: maxChars,
-      signal: opts.signal,
-    });
-  }
-  const timeoutMs = timeoutSec * 1000;
-
-  const spawnOpts: SpawnOptions = {
-    cwd: opts.cwd,
-    shell: false, // no shell-expansion — see header comment
-    windowsHide: true,
-    // PYTHONIOENCODING + PYTHONUTF8 force any spawned Python child
-    // (run_command running `python script.py`, etc.) to emit UTF-8
-    // on stdout/stderr. Without this, Chinese-Windows defaults
-    // Python's stdout encoder to GBK and `print("…")` raises
-    // UnicodeEncodeError on emoji / non-GBK chars — the model then
-    // sees a Python traceback instead of the script's real output
-    // and goes around in circles trying to fix the wrong problem.
-    // Harmless on non-Python processes (env vars they don't read).
-    env: { ...process.env, PYTHONIOENCODING: "utf-8", PYTHONUTF8: "1" },
-  };
-
-  // Windows: two layered fixes on top of shell:false —
-  //   1. Resolve bare command names via PATH × PATHEXT (CreateProcess
-  //      ignores PATHEXT, so `npm` alone misses `npm.cmd`).
-  //   2. Node 21.7.3+ (CVE-2024-27980) refuses to spawn `.cmd`/`.bat`
-  //      directly even with shell:false and safe args — throws
-  //      EINVAL at invocation time. Wrap those via `cmd.exe /d /s /c`
-  //      with verbatim args + manual quoting, so shell metacharacters
-  //      in arguments stay literal.
-  // Unix path is unchanged.
-  const { bin, args, spawnOverrides } = prepareSpawn(argv);
-  const effectiveSpawnOpts = { ...spawnOpts, ...spawnOverrides };
-
-  return await new Promise<RunCommandResult>((resolve, reject) => {
-    let child: import("node:child_process").ChildProcess;
-    try {
-      child = spawn(bin, args, effectiveSpawnOpts);
-    } catch (err) {
-      reject(err);
-      return;
-    }
-    // Collect raw Buffer chunks rather than decoding incrementally —
-    // a multi-byte sequence can land split across chunks, and a naïve
-    // chunk.toString() corrupts it before the second half arrives.
-    // We decode once at close time, where smartDecodeOutput can also
-    // sniff non-UTF-8 codepages cleanly. The byte cap mirrors the
-    // prior char cap (2× maxChars worth) so a chatty process can't
-    // OOM us.
-    const chunks: Buffer[] = [];
-    let totalBytes = 0;
-    const byteCap = maxChars * 2 * 4; // worst-case 4 bytes/char for utf-8/gbk
-    let timedOut = false;
-    let aborted = false;
-    const killChildTree = () => killProcessTree(child);
-    const killTimer = setTimeout(() => {
-      timedOut = true;
-      killChildTree();
-    }, timeoutMs);
-    const onAbort = () => {
-      aborted = true;
-      killChildTree();
-    };
-    // Check synchronously first — if the signal aborted before listener attach
-    // (parent loop was already cancelled), addEventListener with `once:true`
-    // never fires, child runs unbounded.
-    if (opts.signal?.aborted) {
-      onAbort();
-    } else {
-      opts.signal?.addEventListener("abort", onAbort, { once: true });
-    }
-
-    const onData = (chunk: Buffer | string) => {
-      const b = typeof chunk === "string" ? Buffer.from(chunk) : chunk;
-      if (totalBytes >= byteCap) return;
-      const remaining = byteCap - totalBytes;
-      if (b.length > remaining) {
-        chunks.push(b.subarray(0, remaining));
-        totalBytes = byteCap;
-      } else {
-        chunks.push(b);
-        totalBytes += b.length;
-      }
-    };
-    child.stdout?.on("data", onData);
-    child.stderr?.on("data", onData);
-    child.on("error", (err) => {
-      clearTimeout(killTimer);
-      opts.signal?.removeEventListener("abort", onAbort);
-      reject(err);
-    });
-    child.on("close", (code) => {
-      clearTimeout(killTimer);
-      opts.signal?.removeEventListener("abort", onAbort);
-      const merged = Buffer.concat(chunks);
-      const buf = smartDecodeOutput(merged);
-      const output =
-        buf.length > maxChars
-          ? `${buf.slice(0, maxChars)}\n\n[… truncated ${buf.length - maxChars} chars …]`
-          : buf;
-      resolve({ exitCode: code, output, timedOut });
-    });
-  });
-}
-
-/** GBK fallback on Windows — cmd.exe's localized error DLL and native EXE stderr ignore chcp 65001. */
-export function smartDecodeOutput(buf: Buffer): string {
-  if (buf.length === 0) return "";
-  try {
-    return new TextDecoder("utf-8", { fatal: true }).decode(buf);
-  } catch {
-    // Fall through to platform-specific fallback.
-  }
-  if (process.platform === "win32") {
-    try {
-      // TextDecoder supports gbk / gb18030 in Node 18+ via the WHATWG
-      // Encoding spec. gb18030 is the modern superset; falling back
-      // to it covers GBK byte sequences plus the rare 4-byte CJK
-      // characters that appear in newer system messages.
-      return new TextDecoder("gb18030").decode(buf);
-    } catch {
-      // Decoder unavailable in this build — fall through.
-    }
-  }
-  // Last resort: lossy UTF-8 with replacement chars. The model still
-  // gets "something happened" with the structural exit-code marker
-  // intact, which is more useful than throwing away the entire output.
-  return buf.toString("utf8");
-}
-
-export interface ResolveExecutableOptions {
-  platform?: NodeJS.Platform;
-  env?: { PATH?: string; PATHEXT?: string };
-  isFile?: (path: string) => boolean;
-  pathDelimiter?: string;
-}
-
-/** CreateProcess ignores PATHEXT — bare `npm` fails ENOENT under `shell:false` without this resolver. */
-export function resolveExecutable(cmd: string, opts: ResolveExecutableOptions = {}): string {
-  const platform = opts.platform ?? process.platform;
-  if (platform !== "win32") return cmd;
-  if (!cmd) return cmd;
-  // Already a path fragment — spawn handles these natively.
-  if (cmd.includes("/") || cmd.includes("\\") || pathMod.isAbsolute(cmd)) return cmd;
-  // If the model wrote `npm.cmd` explicitly, respect that verbatim.
-  if (pathMod.extname(cmd)) return cmd;
-
-  const env = opts.env ?? process.env;
-  const pathExt = (env.PATHEXT ?? ".COM;.EXE;.BAT;.CMD")
-    .split(";")
-    .map((e) => e.trim())
-    .filter(Boolean);
-  const delimiter = opts.pathDelimiter ?? (platform === "win32" ? ";" : pathMod.delimiter);
-  const pathDirs = (env.PATH ?? "").split(delimiter).filter(Boolean);
-  const isFile = opts.isFile ?? defaultIsFile;
-
-  for (const dir of pathDirs) {
-    for (const ext of pathExt) {
-      // Force win32 join so CI tests that pass `platform: "win32"`
-      // from a Linux runner get backslash-joined paths; the real-
-      // Windows runtime path lands here too and gets the correct
-      // separator regardless of where pathMod defaults.
-      const full = pathMod.win32.join(dir, cmd + ext);
-      if (isFile(full)) return full;
-    }
-  }
-  return cmd;
-}
-
-function defaultIsFile(full: string): boolean {
-  try {
-    return existsSync(full) && statSync(full).isFile();
-  } catch {
-    return false;
-  }
-}
-
-/** Windows workarounds: PATHEXT lookup + CVE-2024-27980 prohibition on direct `.cmd`/`.bat` spawn. */
-export function prepareSpawn(
-  argv: readonly string[],
-  opts: ResolveExecutableOptions = {},
-): { bin: string; args: string[]; spawnOverrides: SpawnOptions } {
-  const head = argv[0] ?? "";
-  const tail = argv.slice(1);
-  const platform = opts.platform ?? process.platform;
-  const resolved = resolveExecutable(head, opts);
-
-  if (platform !== "win32") {
-    return { bin: resolved, args: [...tail], spawnOverrides: {} };
-  }
-
-  // `.cmd` / `.bat` wrappers require cmd.exe on post-CVE Node.
-  if (/\.(cmd|bat)$/i.test(resolved)) {
-    const cmdline = [resolved, ...tail].map(quoteForCmdExe).join(" ");
-    return {
-      bin: "cmd.exe",
-      args: ["/d", "/s", "/c", withUtf8Codepage(cmdline)],
-      // windowsVerbatimArguments prevents Node from re-quoting the /c
-      // payload — we've already composed an exact cmd.exe command
-      // line. Without this Node wraps our already-quoted string in
-      // another round of quotes and cmd.exe can't parse it.
-      spawnOverrides: { windowsVerbatimArguments: true },
-    };
-  }
-
-  // Bare command names that PATH × PATHEXT couldn't resolve to an
-  // on-disk file — these are almost always cmd.exe built-ins (`dir`,
-  // `echo`, `type`, `ver`, `vol`, `where`, `help`, …) which don't
-  // exist as standalone executables. Direct spawn crashes with ENOENT;
-  // routing through cmd.exe lets the built-in resolve, and if it's
-  // genuinely unknown the user gets the standard "'foo' is not
-  // recognized" message instead of a raw spawn failure.
-  if (isBareWindowsName(resolved) && resolved === head) {
-    const cmdline = [head, ...tail].map(quoteForCmdExe).join(" ");
-    return {
-      bin: "cmd.exe",
-      args: ["/d", "/s", "/c", withUtf8Codepage(cmdline)],
-      spawnOverrides: { windowsVerbatimArguments: true },
-    };
-  }
-
-  // PowerShell variants: chcp 65001 doesn't help here because PowerShell
-  // sets its own [Console]::OutputEncoding at startup — usually system
-  // codepage (CP936/CP932/CP949 on CJK Windows) or UTF-16. The result
-  // is mojibake when our `chunk.toString()` UTF-8-decodes its stdout.
-  // Inject a UTF-8 setup prelude into the `-Command` (or `-c`) arg so
-  // any output produced thereafter is UTF-8.
-  if (isPowerShellExe(resolved)) {
-    const patched = injectPowerShellUtf8(tail);
-    if (patched) {
-      return { bin: resolved, args: patched, spawnOverrides: {} };
-    }
-  }
-
-  return { bin: resolved, args: [...tail], spawnOverrides: {} };
-}
-
-/** Resolved bin path looks like Windows PowerShell or PowerShell Core. */
-function isPowerShellExe(resolved: string): boolean {
-  return /(?:^|[\\/])(?:powershell|pwsh)(?:\.exe)?$/i.test(resolved);
-}
-
-/** Targets `-Command` only — PowerShell quoting is finicky enough that wrapping script-file mode could break it. */
-export function injectPowerShellUtf8(args: readonly string[]): string[] | null {
-  const prelude =
-    "[Console]::OutputEncoding=[System.Text.Encoding]::UTF8;$OutputEncoding=[System.Text.Encoding]::UTF8;";
-  for (let i = 0; i < args.length; i++) {
-    const a = args[i] ?? "";
-    if (/^-(?:Command|c)$/i.test(a) && i + 1 < args.length) {
-      const out = [...args];
-      out[i + 1] = `${prelude}${args[i + 1] ?? ""}`;
-      return out;
-    }
-  }
-  return null;
-}
-
-/** Single `&` (not `&&`) so the command still runs on Win7 where chcp can return non-zero. */
-export function withUtf8Codepage(cmdline: string): string {
-  return `chcp 65001 >nul & ${cmdline}`;
-}
-
-function isBareWindowsName(s: string): boolean {
-  if (!s) return false;
-  if (s.includes("/") || s.includes("\\")) return false;
-  if (pathMod.isAbsolute(s)) return false;
-  if (pathMod.extname(s)) return false;
-  return true;
-}
-
-/** Doubles embedded quotes per cmd.exe's `""` escape rule; bare alnum passes through unquoted. */
-export function quoteForCmdExe(arg: string): string {
-  if (arg === "") return '""';
-  if (!/[\s"&|<>^%(),;!]/.test(arg)) return arg;
-  return `"${arg.replace(/"/g, '""')}"`;
 }
 
 /** Error thrown by `run_command` when the command isn't allowlisted. */
